imperium_bot_final/requirements.txt
python-telegram-bot==20.7
aiosqlite==0.19.0
mercadopago==2.2.1
qrcode==7.4.2
python-dotenv==1.0.0
APScheduler==3.10.4
pandas==2.1.4
requests==2.31.0
phonenumbers==8.13.26
validators==0.22.0
cryptography==41.0.8
pytz==2023.3